pyproject.toml
[tool.poetry]
name = "todolistapi"
version = "0.1.0"
description = "Simple Web API for to-do list API"
authors = ["Sikati Kenmogne Samuel <sikatikenmogne@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.8.1"
cfgv = "3.4.0"
distlib = "0.3.9"
django = "5.1.1"
django-filter = "24.3"
djangorestframework = "3.15.2"
filelock = "3.16.1"
identify = "2.6.1"
markdown = "3.7"
nodeenv = "1.9.1"
platformdirs = "4.3.6"
pre-commit = "4.0.1"
pyyaml = "6.0.2"
sqlparse = "0.5.1"
typing-extensions = "4.12.2"
tzdata = "2024.2"
virtualenv = "20.27.0"
sphinx = "^8.1.3"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"
black = "^24.10.0"
isort = "^5.13.2"

[tool.ruff]
exclude = ["todoListAPI/settings.py"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
